Form builder interface and organization (#83)
* Stub out a FormBuilder class

* Merge @atj/documents into the @atj/forms package.
Context: working on FormBuilder, we need to centralize document handling. Documents was prematurely split off, before the interface between it and forms core was worked out. We will likely split off PDF-specific handling into an adapter in the near future, but it will have a different shape than @atj/documents.

* Add setFormSummary() and addDocument() to FormBuilder; wire up to FormManager.

* Add updateFormElement to FormBuilder

* Remove unused arguments from applyPromptResponse