Pattern constructors (#142)
* Fix route parsing in form filler view.

* Create new pattern constructor that validates the config data, and wire up to the external PDF parsing library.

Context: we haven't been validating input from the PDF parsing process, and there were hard to debug issues. A remaining issue that this identified is the presence of spaces in some option IDs. These, and other errors, are currently being logged to the console, and the corresponding pattern is ignored (not added to the form blueprint). Follow-up work should include fixes this issue, and also presentation of these validation errors to the user.

* Split api call and api response processing

* Fix test

* Add mock-response.ts