Add Dockerfile for end-to-end testing
A new Dockerfile, Dockerfile.e2e, has been added to set up end-to-end testing in a docker environment. In this Dockerfile, a container with Node.js and playwright preinstalled is initialized and the project dependencies are installed. The README.md has also been updated to include instructions for building and running this Dockerfile in a docker container.